pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "testmd"
version = "0.1.0"
description = "Manual test tracking driven by TEST.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
testmd = "testmd.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/testmd"]

[tool.pytest.ini_options]
testpaths = ["tests"]
